Plug in to the CLI the lang parsing bits
Python's cmd.Cmd class does it's own parsing of the incoming lines. In order to
use Harmony's lang module, I had to override the parseline method to run the
line through the parser and semantic analyzer.

There are a bunch of stubs in here still. Lots of work left to do.