Finally, a persistence model I'm happy with
Implementation of a SQLiteDatabase object that manages the connection to a
SQLite database, as well as generates appripriate SQL to perform queries. I like
it because it's written in a pretty loose, generic way, so it shouldn't be hard
to swap out this layer if I decide to go with something else. I'm already toying
with ideas for how to mock the backend (though it may be as simple as using an
in memory database) for testing and such.