Model field definitions for SQLite
Okay, so I'm not totally happy with this idea. These two classes define some
options that are specific to how SQLite stores data, but that are really part of
the definition of the models (which is the next layer up). Having these classes
here and separate from the models means I'm duplicating data and will have to
manually keep them in sync. On the other hand, Harmony is small enough that
maybe that's okay. It doesn't really call for a framework as robust as Django...

Still, I think I will have to rethink how to do this.